0.7.2 Updates
Added death loss back in, you now lose about 75% of your items and about 80% of your money on death, equipped items are not lost. Made CH 3 Boss fight mons unrecruitable. Added a book in the player's room that will allow you to change nicknames/team names. This book will be used in the future for other things like changing scarves on the player sprite and for checking statistics.