Musician bug fixes
If you chose a song that plays on a nearby map (i.e. dojo music), the transition going to that map would cause a fade out and back in. I've made it so if the songs match between those maps now because you just REALLY like the guild music or something, it'll stay in and wont fade out and back in.