Chapter 5 work
Midpoint and crucible cutscene work. Fix Appraisal NPC (some of the code was busted, and her animations were not the ones i made for her so long ago). Minor fix to LookAround code thanks to MistressNebula finding its logic before was kinda weird. Fix a potential bug in the rankup item giving code, as amount was not needed and not defined (it's been removed now). Add slugma animation from SE3.